--- src/main/java/edu/handong/csee/isel/parser/ParseType.java
@@ -1,5 +1,0 @@
-package edu.handong.csee.isel.parser;
-
-public enum ParseType {
-	Jira, GitHub, Keywords
-}
