+++ src/main/java/edu/handong/csee/isel/patch/parser/PatchParseType.java
@@ -1,0 +1,5 @@
+package edu.handong.csee.isel.patch.parser;
+
+public enum PatchParseType {
+	Jira, GitHub, Keywords
+}
