+++ src/main/java/edu/handong/csee/isel/patch/parser/githubparser/NoIssuePagesException.java
@@ -1,4 +1,4 @@
-package edu.handong.csee.isel.utils;
+package edu.handong.csee.isel.patch.parser.githubparser;
 
 public class NoIssuePagesException extends Exception {
 	String message;
